test: moved python file to utils
